pyproject.toml
[tool.poetry]
name = "proxy-list"
version = "0.1.0"
description = ""
authors = ["Вадим Акимов <sfc.coltrane@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14"
celery = "^5.3.6"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
requests = "^2.31.0"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
